10165 FIX Factory settings: Docker container host check command is set now
Instead of always assuming the docker containers to be up, the docker containers
now base their host state on the service state of the "Docker container status"
service.

This is configured by the factory settings using the ruleset "Host Check Command".
The factory settings only apply to new created sites. In case you want this in
existing sites, you need to create a rule in this ruleset manually.

Change-Id: Ibf4fd91a46b9d17187e396b834ea65ea28678bcf