7278 FIX cmk-update-config: fixed exception during ruleset conversion
The ruleset conversion from 1.5 and earlier versions to 1.6 randomly failed.
In this scenario the "cmk-update-config" command showed an exception on the command line.
This has been fixed.

Note: If you want to the config update to complete in the previous version, you
can call the <tt>cmk-update-config</tt> command repeatedly until it finally runs through.

Change-Id: Ieb834998d17d83d5d095dfdf7b5a6c4c493c3eaa