10179 FIX Discovery page: Missing "full scan" button in some cases
In previous 1.6 releases the message "No discovery information available.
Please perform a full scan." could appear without a visible "Full scan"
button could appear making it hard to proceed with a service discovery.

As a workaround with previou versions, you could execute a bulk discovery for
this host or go on the diagnostic page and click on the "Test" button once.
Afterwards you should be able to proceed on the discovery page.

Change-Id: I0b05fff29cf03022cdff4e8a698c11417e3e201c