10071 graylog_failures: New check to monitor the number of failures across all indices
With this check you can monitor the number of failures across all indices of a
graylog instance. It outputs the total number of failures and the number of
failures in a given timeframe (default: last 30m). Also the number of affected
indices of the last failures are shown. The long output shows detailed
information of the last 30 failures.

Change-Id: I65a747248aca58213c4e9a7ae3a991ac215c7521