Start agent integration tests + Add python plugin compatibility tests
Start new type of tests: agent-integration. These are meant to execute
agents or agent plugins in a specific environment.

The first tests are meant to test the compatibility of python plugins
with different python versions. For the moment they simply check for
syntax compatibility. We may execute the unit tests in the near future
in these environments.

Change-Id: If85365ae84ad1662ba7c1001029a7ad0dce24830