10142 mk_logwatch: Explicitly specify encoding of logfiles
mk_logwatch tries its best to determine the correct encoding of a file,
and to recover from errors if it is wrong.
If that does not work for you, you can configure a specific encoding using
the bakery rule or by adding e.g. "encoding=latin1" to the options in the
config file (see config examples in {{~MySITE/share/check_mk/agents/cfg_examples/}}).

Change-Id: I88d87ae2ae71e2953edfd88052f0771d3d25503c