10184 FIX mk_docker mk_logwatch mtr mk_jolokia apache_status: Fixed Python 2.5 syntax compatibility of agent plugins
The agent plugins need to be executed on older Linux systems which may
have very old Python versions available. For this reason we need to use
old Python 2 syntax here.

Python plugins that are incompatible to 2.5, for example because some
3rd party library is not available with 2.5, need to be syntax
compatible with 2.5 for the moment, but are allowed to terminate with a
helpful error message about this incompatibility.

Change-Id: I52249653c1046412a5de543d143673b884a8d637