10267 FIX dell_complellent_disks would not see more than 9 disks
The dell_compellent_disks check was using the the inventory function
from `dell_compellent.include`.  However, this inventory function
expects a list where `parse_dell_compellent_disks` builds a dict.  This
resulted in the inventory function returning the first letter of every
key from the parse function, that is, disk #7, would be correctly
inventoried as #7 but disk #42 would be inventoried as #4.  The
inventory function would also never see more than 9 disks with no
guarantee that it would always see the same 9 disks.

This patch fixes the inventory function to dell_compellent_disks
and test that the error is not present in the other Compellent checks.

This fixes FEED-4283.

Change-Id: I16cfd52f220b7cc2625443ed8bf4f1ee2a748451