10244 FIX Periodic service discovery was not correctly activating changes
When the automatic update of the service configuration is enabled, the periodic
service discovery was applying service discovery changes, but was not
activating them when configured to do so.

The services appeared as discovered on the discovery page in WATO, but they were
not activated. After activating a change in the GUI, the services were finally
activated.

Change-Id: Ic4902ceb7a2fe1fb216ed438ed40b30417d8eef9