Fixed handling of stderr in local automation calls.
Merging stdout and stderr was totally wrong: The former will be parsed as a
Python literal, the latter (if non-empty) is textual message for the user.
Alas, automation calls have no nice way to transfer the message from stderr
to their call sites, so we just log them to var/log/web.log.

Change-Id: I1728433813dc6095429c1c2a39a522614a0b6e8d