10252 FIX Fix broken "Start livestatus proxy" global setting after update from CRE to CEE
The global setting "Start livestatus proxy" is used to enable the livestatus proxy
daemon in Checkmk sites. It is a "omd config" setting which can also be set using
the Checkmk GUI.

Previous 1.6 versions did not correctly initialize new OMD configuration settings
after "omd update". When updating from CRE to CEE the new configuration option
<tt>CONFIG_LIVEPROXYD='on'</tt> should be added to <tt>etc/omd/site.conf</tt>,
because this file needs to hold the current effective config options for all
available OMD settings. The missing update of this file resulted in a
"KeyError ('site_liveproxyd')" exception in the GUI when trying to edit this
setting.

Change-Id: I8efc9c639f8d38706f71dcd994eb5449ab83b8c5