10323 graylog_sidecars: New check to monitor sidecars of graylog
With this check you can monitor sidecars of a graylog instance. It outputs the
state of the sidecar, when the sidecar was last seen and the state of the
collectors of the sidecar. Collector details can be found in the long output of
the service.

Needs the datasource "Check state of Graylog" to be configured.

Change-Id: Ifb062b0d07058a27fc4d48aa9d1bbde68b65d888