10153 jolokia_jvm_garbagecollectors: New plugin
The Services "JVM MyInstance GC MyGarbageCollector" have been moved to a dedicated Plugin.
Users of the special agent Jolokia or the latest version of the mk_jolokia
plugin must perform a rediscovery.
The performance metrics are kept and defined rulesets will apply to the new
service (by the same name).

Change-Id: I3b504b5c077fd7bad02069c2bf8ebee1f1d86ac0