Fixed handling of stderr in local automation calls.
Merging stdout and stderr was totally wrong: The former will be parsed as a
Python literal, the latter (if non-empty) is textual message for the user.
Alas, automation calls have no nice way to transfer the message from stderr
to their call sites, so we just log them to var/log/web.log.

Change-Id: I4aa098b27bdc4daab6bbf2afe993b0e418689bf9