10260 FIX stunnel service is now correctly disabled when LIVESTATUS_TCP_TLS is off
The stunnel daemon of the site, that is used for serving the encrypted
livestatus communication was activated just by enabling the LIVESTATUS_TCP,
even when the "omd config" option LIVESTATUS_TCP_TLS was set to off. This was
not a real problem from a user perspective because that stunnel process was
inactive in this situation.

Change-Id: I2e67c578eec21731ce5fd4d73d2681fcb0e60605