08796 FIX mknotifyd: fix wrong state file age output
The computation of the state file age was dependent on the
execution time of the check_mk_agent and the scheduling of
the check itself. If e.g. the sections after the mknotifyd
section took 5s this time was added to the state file age
inside the check. In the worst case this may lead to wrong
warnings that the state file age is older than 90s.

To use the fixed version of the mknotifyd check the
check_mk_agent has to be updated on systems with Checkmk
installations after a site update.

CMK-2897

Change-Id: I2f90b48f49856e7c58de1d9ee8cf8fb95b1111bb