10261 FIX API: The edit_users call can now be used to edit LDAP users
The web API call "edit_users" that can be used to edit the attributes of a user
could previously not be used for users that were created using the LDAP synchronization.
This limitation has now been removed.

Please note that the call can only set attributes that are not locked by a LDAP attribute
sync plugin (see LDAP connection settings). This works like the "Edit user" dialog.

Change-Id: I11165e2ae58bd5ffc2349d4ffb10435016d7097b