10307 FIX Resolve conflict event console archive event
There are two methods to archive events in the event console view. By
single event, directly clicking on the archive icon or by bulk selection of
rows. In the case the checkboxes for row selection were toggled by default
the single event case was shadowed and could not find the selected event.

With Werk resolves the issue and allow both actions to play together.

Change-Id: I066cbc3e2f01225d518e22c487f09aeb786ac3a1