CMK-3011: cmk-update-agent.exe: Show pending update in service
Only relevant for calls of cmk-update-agent.exe as a plugin:

A pending update (i.e. agent package is placed at installation)
folder and the agent updater awaits the update by Windows Agent)
is now mentioned in the AgentUpdate agent section and processed
by the check_mk_agent check.
Under normal circumstances, it will barely visible on the GUI
because the Windows Agent is very quick with the update and the
agent updater will be called immediately after the update.
But if the Windows Agent fails to update, this would be a clue.

Additionally, the last_check value is now set more reasonably:
It is now only updated if an actual check for an update took place.
(Not if just the pending update gets confirmed)

Finally, the agent updater now pushes its status to the server even
if it is only awaiting the Windows Agent Update, because this marks
the "last contact" at the agent updates status page.

Change-Id: I7f9b5f3d93495493d22276e57234e64914a94688