8798 FIX agent_kubernetes: allow the option 'No IP' for the Kubernetes master
Previously it was necessary to specify an IP for the Kubernetes master.
If the option 'No IP' was used, the special agent was not able to connect
to the master even when a 'Custom URL prefix' was given. Now the IP is
not necessary anymore when a URL prefix is given.

Change-Id: Ic0c0b075d625290afdb613210a56461be3ebe9ca