10309 FIX Predictive Levels is robust against receiving no data at all
On a new site, when perfdata is not available and predictive levels are
set. Livestatus replies with [0, 0, 0] -> step==0, and a zero division
Error is raised leaving Check in UNKN and no new perfdata is delivered,
which makes the problem permanent. Extra flow control on retrieved data
allows for the predictive levels not to be calculated until there is enough
data available. The check now continues to output such data.

Change-Id: I2778e4929b86476db214ba48b5603e0acd662cb7