Helper: track import memory
This program overwrites __import__ so that we can measure
where in the import-graph the most memory is being used.

It can emit the data in various formats. For now graphviz
and a json-tree structure are supported.

It works completely stand alone and (with the exception being
distutils) has no further dependencies than the standard
library.

Change-Id: I436f6e4b140c08e16d4b6431b39ee33a95b57178