10159 New plugin: Couchbase nodes info
A new plugin is added which monitors the health status of a cluster node.
The service states for "warmup_state" and "unhealthy_state"
can be configured in the WATO rule "Couchbase status".
By default the service is CRIT on "unhealthy_state"
and OK on any other state.

It additionally displays whether it is a OTP node, the recovery type,
the version and the cluster compatibility.

It also monitors the cluster membership state, counfigurable in the rule.
By default inactive added notes go to WARN and inactive failed notes to CRIT.

Change-Id: I52bd326655ef52fe7adf11b28353e3971ccdba29