10194 Windows Agent: logwatch section size is limited now
Previously the Windows Agent may generate too much data during Windows EventLog
(logwatch) section generation. In fact, size of the data to be send was
unlimited, the Agent could have created a block of data several gigabytes
in the size.

Now default maximal size of the whole logwatch section of the Windows Agent
is limited to 500000 bytes. The value of this limit may be changed in
check_mk.user.yml using variable logwatch.max_size. To have unlimited
section, the value of the max_size should be set in -1

Change-Id: I62d7f036c16a53ef365152523959f45293876d97