10195 Windows Agent: additional parameters for logwatch section
To better control output from the logwatch section the user can set next
parameters for this section:

max_line_length, max_entries, timeout

timeout is measured in seconds and limits the  time of the execution of the
every reported log.
max_line_length cuts every line in output data if the size of the line
exceeds the provided value.
max_entries is the maximal count of the entries in every reported event log.

The default values for all above mentioned parameters are -1 and means
unlimited.

Change-Id: Iabcd177f4cf514b3201b7695e0c7846c617b63a5