events.py: use logging module
Introduce the logging module in events.py. This
makes it possible to remove the log_function argument
in event_keepalive and the log_func argument in
complete_raw_context. Finally, the alert_log function
in alert_handling.py and the notify_log function in
notify.py can be removed.

CMK-1657

Change-Id: I7a7cd4e6f5b8f446a6264cecc9b8dffc5378ed5f