10427 cmk-update-agent: Show info about non-existing host on registration
On agent registration, an information will now be displayed if the hostname
that just got registered is not known in the activated monitoring configuration.

Please note that the spelling of the registered host and the monitored host
must match exactly to enable automatic updates.

The displayed information should help to avoid errors and to clarify things.

CMK-1120

Change-Id: Iba89f1ee631852a3d24c007fdc36a980ee698a9b