Initial travis-ci configuration
This is a simple entry point to execute the basic and most important
Python tests for Checkmk. We run tools like pylint, YAPF and our py.test
based unit tests here. Some tests, like integration tests or tests of
very specific components are not executed.

We focus on the tests that are needed by external developers, e.g. to
support check plugin development.