10449 FIX ps: cleanup counters of processes which do not exist anymore
The new section ps_lnx was introduced in version 1.6.0. Unfortunately,
counters of processes which do not exist anymore were not cleaned
up for this section. This may lead to slowly growing files in
tmp/check_mk/counters for hosts with the ps_lnx section. Since the
check helpers have to parse this files on a regular basis this leads
to an increased CPU load.

It's worth to note that counters only accumulate if you use process
monitoring and only if the PIDs of these processes change frequently.
If you do not use process monitoring or the PIDs of your monitored
processes do not change, you are not affected by this issue.

CMK-3090

Change-Id: I6a2b68e75b03acf5f04c9d1fa2ea3814b3511de0