Simplify dpkg-sig call
Turns out most of the options are not relevant, because they are set
by dpkg-sig.

Since w only sign one file per dpkg-sig call it does not seem to be
needed to use the "cache passphrase" option. The additional logic
in dpkg-sig might even cause the random issues we experience from
time to time.

In case it does not help we could increase the log levels of dpkg-sig
and gpg to get more information (--verbose -g -v).

Change-Id: I5f906e225350d1b83e123151d011075e41969539