10290 FIX Temperature check: fix rendering crash and reporting mismatch
When a check sends a device-specific levels in a specific format
the temperature rendering function of the service display
could crash due to a type mismatch. This werk fixes this crash.

Upon fixing this crash another bug was discovered and fixed. In the case
of user specified levels being set and the service being in mode
"Use your own levels if present, otherwise the device's", the device
levels were not ignored and thus the wrong levels were checked. This
is now fixed as well.

No additional action is needed.

Change-Id: Ia5a3be37b657955fecef8df48c887c9c23f7fcbf