10404 FIX Performace update for RRD on high resolution
When the resolution of RRD files is increased it could lead to potential
delays in rendering of the graphs. This werk separates the tasks of
flushing the RRD files and querying them for data, which leads to a
performance increase, because the daemon is not called by the xport tool.

The performance issues with the cache daemon and xport tool have been
reported to RRDTool on the issue:

https://github.com/oetiker/rrdtool-1.x/issues/1062

CMK-1891

Change-Id: I547face0ae7f9badd1498be20b2f2a3fa818858f