10383 FIX Make more background job results deletable
The results of some background job types could not be deleted using the GUI.
For example the "User synchronization" or "Agent baking" results can now be
removed on the Background job overview page. Job results executed in versions
previous to 1.6.0p6 will not be deletable, only the results of jobs executed
with 1.6.0p6 or newer.

If you want to delete the jobs, you can remove the job result folders from
<tt>var/check_mk/background_jobs/[job-type]</tt> manually.

Change-Id: I0d3f01154ea73cc42a581189929904f213ff4ac6