10330 graylog_license: New check to monitor state of Graylog license
This check monitors the state of the license used with a Graylog instance. It
outputs the expired, violated and valid state, if the traffic limit is
exceeded, if the cluster is not covered and if nodes exceeded. Furthermore it
reports if remote checks failed, the traffic limit, the time until the license
expires and informational, the subject of the license, if license is a trial
and if it requires remote checks.

Needs the datasource "Check state of Graylog" to be configured.

Change-Id: I93486fbc7e4e9c9eb4591a4cdea349e529564dc2