10471 mk_logwatch: Add option to limit context
If the plugin mk_logwatch is configured to send context along with found messages,
the amount of data can become quite large. This werk adds the option of limiting
the context given for every warning or critical message to a given number of lines
befor and after the message. For instance, to limit the context to 3 lines before
and four lines after the message, set the option "maxcontextlines=3,4".

This option can be configured using the WATO rule "Text logfiles".

CMK-2923

Change-Id: I8145b348fd8e466ad182f878108664c311f462d9