10437 emc_ecs_mem: New SNMP memory check for Dell EMC ECS storage systems
New SNMP check for monitoring the memory status for Dell EMC ECS (cloud storage system). Levels for the memory usage can be set.
The system already delivers levels for the minimum swap space that needs to be available.
Make sure to create the rule 'Hosts without system description OID' for this host.

Change-Id: If65e20d543f0e7bdc6b31c0876021ef0274afcbb