check_http: don't use the -H option when a proxy is used
In the proxy case the hostname is not set via the -H option,
but given as the second positional argument. If -H is given
anyway the second positional argument will override it. To
make the use of the second positional argument more explicit
we now omit the -H option entirely in this case.

Add a test case along the way.

CMK-3184

Change-Id: Ib241d2beb4fe979cd8b054fab2d927f35ac03332