10451 FIX check_http: don't set virtual host if not necessary
Since version 1.6 the active check check_http automatically
set the HTTP Host header and used HTTP/1.1 instead of HTTP/1.0
even if no virtual host was supplied. This may lead to
timeouts of the check_http plugin on servers which are not
fully compliant with HTTP/1.1.

This werk changes this behaviour so that the Host header is
only set if a virtual host is given.

Please note that if you rely on HTTP/1.1 or want to use it you
have to supply a virtual host.

Please also note that the virtual host will not be set
automatically to the host name anymore. If you relied on this
behaviour you now have to provide the virtual host in the
affected rules.

CMK-3184

Change-Id: I82c1936acc4d4cba8265b21f6a5b587e9823c9f4