10408 FIX Don't discover lparstat service on host without util info
After update to 1.6 you might find a lparstat service with UNKN status on
your AIX hosts. This service should have not been discovered as no data is
available to run the check plugin on it.

You should delete that service or perfom a rediscovery after this werk in
case you have that service.

CMK-3106

Change-Id: Iec27d843b58921b071b1d9aa800a040bb365d5ac