get the table offset with interface index
in logically separated Brocade FC switches the interface indices do not start with 1 and therefor the index in the speed table does not start with 1073741824.

It is necessary to add the first interface index as offset to find the table offset in the speed table.

This is a better fix than PR #29 https://github.com/tribe29/checkmk/pull/29