Re-add rrd stuff to the os containers
These are used for our integration test, should not compile anything on
their own, but are currently executing our integration tests, which need
the standard virtual environment, which includes the rrd python binding.

We could now add more compilation dependencies to our os containers and
make the rrdtool python build with our own python or skip the
compilation and make the virtual environment use the rrdtool library of
the system.

For the moment we use the later approach. For the future it may be
better to reduce the needs in the container.

Change-Id: I7c81a2ae24fc543a1c53cb29e8645bb4548a1ed7