10550 FIX Fix moving table rows (e.g. EC rules) when table action row is shown
When trying to move Event console rules in the table while the action row
at the top of the table is enabled, the rows were always positioned one
row below the position you wanted to put it to. This could be solved by
temporarily hiding the action row before moving a rule.

The dragging is now working even when the action row is enabled.

Change-Id: I51455346684f639d1def9b2c5670b22f80dc37ab