10460 Show an error in the tactical overview if sites are not connected
If a remote site is not connected in a distributed setup the hosts,
services and events of that site are not taken into account in the
tactical overview of the master site. To indicate that information
is missing the tactical overview now displays an error if sites are
not connected. The error is displayed below the rows of the tactical
overview similar to failed notifiactions.

CMK-3147

Change-Id: I92aad8a8edbb6a240e52b7a7d7777ebedc241231