10200 FIX WMI timeout in 'ps' section doesn't prevent Windows Agent output
Previously, if a timeout occurs in the 'ps' section during WMI query, the agent
immediately stopped generation of data from all remaining sections. As result,
the monitoring site reported a lots of missing and/or staling services.

Now, in the case of the timeout in the 'ps' section the Agent skips only this
section.

Change-Id: I44fdd129d96539bff532e14e1306f3209d3fdd17