Do not use the magic method __fspath__ directly.
Magic methods are, well, magic, so they should normally not be called
directly.  The official Python 3 way for

    foo.__fspath__()

is

    os.fspath(foo)

Alas, Python 2 doesn't have os.fspath(), but in our case we are lucky: If we
know that foo is a (Pure)Path, we can simply use str(foo), which is the
equivalent for Python 2 *and* 3. (Note the different return types!)

Change-Id: I5090187e0bb1a18143d18a8d4aae8d814dc06f10