Make rrddata available to Nagios
The `*RRDColumn` classes used to call directly in CMC.  Moving them
implies to trim anything related to CMC off the class and put it
into subclasses.  So that after refactoring we have:

 +----------------------+               +---------------+
 | DynamicRRDColumn (A) | ------------> | RRDColumn (A) |
 +----------------------+               +---------------+
     ^                                          ^
    / \                                        / \
    ---                                        ---
     |                                          |
     |                                          |
     |                                          |
  +--------------------------------------+      |
  | Dynamic{Object,Host,Server}RRDColumn |      |
  +--------------------------------------+      |
                                                |
                                +-------------------------------+
                                | {Object,Host,Server}RRDColumn |
                                +-------------------------------+

Where the concrete `*ObjecRRDColumn` classes only appear in CMC
while the `*{Host,Server}RRDColumn` classes only appear
in the Nagios core.

This is different from, for example, the `{Host,Service}ContactsColumn`,
classes where the behavior of the CMC is not encapsulated in its own
class but separated from the Nagios code by macros.

This closes CMK-3269.

Change-Id: I728e61fb164c70420f2bcd4fccb98172f15f548d