Rework package cache
The previous approach to hook between the build and install phase did
not work very well because build specific things like absolute paths in
generated build files (Makefiles and so on), which can be system
dependent, can not be cached. Packing the files after build on one
system, unpacking it on the other system and then trying to perform the
install steps failed e.g. for Python because of absolute paths in the
cached files.

The new approach is to hook into the installation phase. The idea is now
that every package now installs it's files in a package individual
directory which contains the files the package wants to install.  The
files need to be relative to $OMD_ROOT, for example
$INTERMEDIATE_INSTALL_BASE/Python/bin/python will finally result in a
file /opt/omd/versions/[version]/bin/python.

This intermediate install result is cached and uploaded to or downloaded
from the cache. After that is done all package installed files will be
merged together into the final install destination.

Change-Id: I29f7c7a6ec6fad51d3974428aeadab2eb0f3d795