Brocade_port: Fix priority in if_table for recognizing speed
Brocade speed :swFCPortSpeed is deprecated in firmware 7.2.1 as for comment
is check. But we only select the new speed value for fibre channel.

The if_table had a if clause in the list comprehension with wrong operator
priority.

Brocade speed maps from a dictionary with integer keys, it never recognizes
str values. It was priority, but now that is fixed we need to skip that
return for the fibre channel new speed.

CMK-3187

Change-Id: Iac998edfb420742b9ecbeb149e62bf6754b1267b