deb: Cleanup issues with DESTDIR set while executing build steps
In the past the debian packaging mechanism in debian/rules
was calling the omd Makefile once for building and once for
installation.

For our dependency logic it is easier to only call the Makefile once
with the install target and make them resolve the dependencies on their
own using the Makefile dependencies. The major difference are the set
DESTDIR and MAKEFLAGS which confuse some "make install" calls. These
calls now unset the variables as needed.

Add some missing mkdir calls on the way.

Change-Id: I399facdf9a27cafcbcfa16a58dd9d801897b41be